# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: FedRAMP Gap Analysis Agent API
  version: 1.0.0
  description: |
    AI-powered FedRAMP High baseline compliance gap analysis agent that analyzes policy documents, 
    design documentation, and source code to identify compliance gaps and provide remediation recommendations.

    ## Features
    - Comprehensive gap analysis across policy, design, and code
    - Support for 325+ FedRAMP High baseline controls (NIST 800-53 Rev 5)
    - Risk assessment with severity classification
    - Automated remediation recommendations with code examples
    - Multiple report formats (JSON, PDF, HTML, Excel)

    ## Authentication
    All API endpoints require authentication using Bearer token or API key.

  contact:
    name: API Support
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.fedramp-agent.example.com/api/v1
    description: Production server
  - url: https://staging-api.fedramp-agent.example.com/api/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Local development server
...
paths:
...
  /reports/{jobId}:
    get:
      tags:
        - Reports
      summary: Get analysis report
      description: Retrieve analysis report in specified format
      operationId: getReport
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [json, pdf, html, excel]
            default: json
          description: Report format
      responses:
        "200":
          description: Report retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResultResponse"
            application/pdf:
              schema:
                type: string
                format: binary
            text/html:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - BearerAuth: []
  /reports/{jobId}/summary:
    get:
      tags:
        - Reports
      summary: Get report summary
      description: Get executive summary of analysis results
      operationId: getReportSummary
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Summary retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  summary:
                    $ref: "#/components/schemas/AnalysisSummary"
                  control_coverage:
                    $ref: "#/components/schemas/ControlCoverage"
                  top_risks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Gap"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    AnalysisResultResponse:
      type: object
      required:
        - job_id
        - analysis_date
        - gaps
        - risk_assessments
        - remediations
        - summary
        - control_coverage
      properties:
        job_id:
          type: string
        analysis_date:
          type: string
          format: date-time
        gaps:
          type: array
          items:
            $ref: "#/components/schemas/Gap"
        risk_assessments:
          type: array
          items:
            $ref: "#/components/schemas/RiskAssessment"
        remediations:
          type: array
          items:
            $ref: "#/components/schemas/Remediation"
        summary:
          $ref: "#/components/schemas/AnalysisSummary"
        control_coverage:
          $ref: "#/components/schemas/ControlCoverage"
    Gap:
      type: object
      properties:
        gap_id:
          type: string
        control_id:
          type: string
        control_name:
          type: string
        gap_type:
          type: string
          enum:
            [
              missing_in_design,
              missing_in_code,
              missing_implementation,
              partial_implementation,
              incorrect_implementation,
              outdated_implementation,
              configuration_gap,
            ]
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        description:
          type: string
        policy_requirement:
          type: string
        design_specification:
          type: string
          nullable: true
        code_implementation:
          type: string
          nullable: true
        evidence:
          type: array
          items:
            type: object
        risk_score:
          type: number
          minimum: 0
          maximum: 10
        impact:
          type: string
        likelihood:
          type: string
    RiskAssessment:
      type: object
      properties:
        gap_id:
          type: string
        control_id:
          type: string
        risk_score:
          type: number
        impact:
          type: string
        likelihood:
          type: string
        risk_level:
          type: string
        business_impact:
          type: string
        technical_impact:
          type: string
        compliance_impact:
          type: string
        exploitability:
          type: number
        remediation_priority:
          type: integer
    Remediation:
      type: object
      properties:
        gap_id:
          type: string
        control_id:
          type: string
        summary:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/RemediationStep"
        effort_estimate:
          type: string
        estimated_hours:
          type: integer
        required_skills:
          type: array
          items:
            type: string
        testing_guidance:
          type: string
        validation_criteria:
          type: array
          items:
            type: string
    RemediationStep:
      type: object
      properties:
        step_number:
          type: integer
        description:
          type: string
        technical_details:
          type: string
        code_example:
          type: string
          nullable: true
        verification:
          type: string
    AnalysisSummary:
      type: object
      properties:
        total_controls_evaluated:
          type: integer
        controls_with_gaps:
          type: integer
        controls_compliant:
          type: integer
        total_gaps:
          type: integer
        critical_gaps:
          type: integer
        high_gaps:
          type: integer
        medium_gaps:
          type: integer
        low_gaps:
          type: integer
        average_risk_score:
          type: number
        compliance_score:
          type: number
    ControlCoverage:
      type: object
      properties:
        total_required:
          type: integer
        total_implemented:
          type: integer
        coverage_percentage:
          type: number
        by_family:
          type: object
          additionalProperties:
            type: object
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...